SoleViz

Welcome to the documentation for SoleViz.

SoleViz._descriptors2grouped

SoleViz._framedims

SoleViz._get_l

SoleViz._get_r

SoleViz._get_t

SoleViz._grouped2descriptors

SoleViz._nmodalities

SoleViz._nwindows

SoleViz.plotdescription

SoleViz._descriptors2grouped — Method

_descriptors2grouped(descriptors)

Convert a Vector of descriptors to its grouped version generating singleton groups.

SoleViz._framedims — Method

_framedims(d)

Get the dimension of each frame in d.

SoleViz._get_l — Method

_get_l(d)

SoleViz._get_r — Method

_get_r(d)

SoleViz._get_t — Method

_get_t(d)

SoleViz._grouped2descriptors — Method

_grouped2descriptors(grouped_descriptors)

Convert a Dict of grouped_descriptors to vector of them.

SoleViz._nmodalities — Method

_nmodalities(d)

Get the number of modalities in d.

SoleViz._nwindows — Method

_nwindows(d)

Get the number of windows d contains.

This is the number of t values, not the number of windows the dimensional data will be divided into.

SoleViz.plotdescription — Method

TODO: docs

TODOs

Support more than one descriptor in descriptors

Support more than one function in functions

Support variable_order to order variables in plots

Implement plot_dimension support 3D visualizations for windows

Add more user-friendly interface for windows parameter